Report cheating
Creation of an item ("Cheater melden") in menu.
Added a specific picture for that item.
Added a "Yes/No" dialog box for reporting the cheater (dialog box will appear after clicking on item).
Notify all players after choosing "Yes".
Condition for not notifying the Mr.X or the player that reported cheating.
Players will receive a toast pop up with the name of a player that reported.