Implemented ENdgame and CHeating
End Game: If one detective and Mrx are on the same position the game ends and the EngGame Activity is shown with the apprpriate MEssage.

Cheating: If MrX is the Server, cheating is implemented. Which means by triggering the proximitry sensor he can make an additional move without letting the client notice. A message will be shown on the clients with a certain probability .