Logic for cheater reporting
Implemented the logic for reporting the Mr. X depending on if he really cheated or not.

In case that a detective reported Mr. X 3 times correctly during the game, Mr. X will lose the game and the detectives will win.

In case that a detective reported Mr. X falsely, the number of his remaining tickets will be randomly decreased. (The visual represantation of tickets being taken away from a detective is still in progress!!).